refactor: GNB 컴포넌트 분리
